# Conflicts:
#	takin-web-biz-service/src/main/java/io/shulie/takin/web/biz/service/LinkTopologyService.java
#	takin-web-biz-service/src/main/java/io/shulie/takin/web/biz/service/report/impl/ReportLocalServiceImpl.java
#	takin-web-biz-service/src/main/java/io/shulie/takin/web/biz/service/report/impl/SummaryService.java
#	takin-web-biz-service/src/main/java/io/shulie/takin/web/biz/service/risk/impl/ProblemAnalysisServiceImpl.java
#	takin-web-data/src/main/java/com/pamirs/takin/entity/domain/entity/report/TpsTarget.java

+++ takin-web-biz-service/src/main/java/io/shulie/takin/web/biz/pojo/output/report/ReportAppInstancePerformanceOut.java
@@ -23,7 +23,7 @@ public class ReportAppInstancePerformanceOut {
     @ApiModelProperty(value = "平均网络使用率")
     private BigDecimal avgNetUsageRate;
     @ApiModelProperty(value = "GC次数")
-    private int gcCount;
+    private BigDecimal gcCount;
     @ApiModelProperty(value = "GC时间")
     private BigDecimal gcCost;
     @ApiModelProperty(value = "平均TPS")
