+++ takin-web-biz-service/src/main/java/io/shulie/takin/web/biz/service/report/ReportMessageService.java
@@ -14,5 +14,5 @@ public interface ReportMessageService {
 
     ReportMessageDetailDTO getOneTraceDetail(ReportMessageDetailReq req);
 
-    Integer getRequestCountByCost(ReportCostTrendQueryReq req);
+    Long getRequestCountByCost(ReportCostTrendQueryReq req);
 }
